requirements.txt
pandas>=2.2.0
numpy>=1.26.0
faker>=20.1.0
pyspark>=3.5.0
matplotlib>=3.8.0
plotly>=5.17.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
python-dotenv>=0.21.0 